v2/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.3.2
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dotenv==1.2.2
typing_extensions==4.15.0
Werkzeug==3.1.8
